+++ src/library/PromptbookLibrary.ts
@@ -1,3 +1,4 @@
+import { Promisable } from 'type-fest';
 import type { Prompt } from '../types/Prompt';
 import type { PromptbookJson } from '../types/PromptbookJson/PromptbookJson';
 import type { string_promptbook_url } from '../types/typeAliases';
@@ -18,7 +19,7 @@ export type PromptbookLibrary = {
      *
      * Note: This is not a direct fetching from the URL, but a lookup in the library
      */
-    getPromptbookByUrl(url: string_promptbook_url): PromptbookJson;
+    getPromptbookByUrl(url: string_promptbook_url): Promisable<PromptbookJson>;
 
     /**
      * Checks whether given prompt was defined in any promptbook in the library