+++ src/errors/NotFoundError.ts
@@ -1,0 +1,10 @@
+/**
+ * This error !!!
+ */
+export class NotFoundError extends Error {
+    public readonly name = 'NotFoundError';
+    public constructor(message: string) {
+        super(message);
+        Object.setPrototypeOf(this, NotFoundError.prototype);
+    }
+}